+++ CHANGES.rst
@@ -12,6 +12,7 @@ Unreleased
 -   Add type annotations.
 -   Adding the CLI is skipped if ``command_name`` is empty rather than ``False``.
 -   The internal cache only holds a weak reference to the Flask app.
+-   Various arguments no longer use a default value when passed ``None``.
 
 
 Version 2.0.1